flatten event,thing,occurences selectors
