+++ src/selectors/occurences.js
@@ -1,25 +1,33 @@
+import get from 'lodash.get'
 import moment from 'moment'
 import { createSelector } from 'reselect'
 
-export default createSelector(
+export default () => createSelector(
   state => state.data.eventOccurences,
-  eventOccurences => {
+  (state, ownProps) => get(ownProps, 'occasion.eventId'),
+  (eventOccurences, eventId) => {
     if (!eventOccurences) {
       return
     }
 
+    let filteredOccurences
+    if (eventId) {
+      filteredOccurences = eventOccurences.filter(eventOccurence =>
+        eventOccurence.eventId === eventId)
+    } else {
+      filteredOccurences = [...eventOccurences]
+    }
+
     // clone
-    const filteredOccurences = [...eventOccurences].map(eo =>
+    filteredOccurences = filteredOccurences.map(eo =>
       Object.assign({}, eo))
 
     // sort by dates
-    if (filteredOccurences) {
-      filteredOccurences.forEach(o => {
-        o.beginningDatetimeMoment = moment(o.beginningDatetime)
-      })
-      filteredOccurences.sort((o1,o2) =>
-        o1.beginningDatetimeMoment - o2.beginningDatetimeMoment)
-    }
+    filteredOccurences.forEach(o => {
+      o.beginningDatetimeMoment = moment(o.beginningDatetime)
+    })
+    filteredOccurences.sort((o1,o2) =>
+      o1.beginningDatetimeMoment - o2.beginningDatetimeMoment)
 
     // return
     return filteredOccurences
